Add google sign in to the automotive app
