+++ modules/features/account/src/main/java/au/com/shiftyjelly/pocketcasts/account/onboarding/OnboardingLoginOrSignUpPage.kt
@@ -2,12 +2,8 @@ package au.com.shiftyjelly.pocketcasts.account.onboarding
 
 import android.content.res.Configuration
 import android.content.res.Configuration.ORIENTATION_LANDSCAPE
-import android.widget.Toast
 import androidx.activity.compose.BackHandler
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.DrawableRes
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -33,7 +29,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalView
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
@@ -42,6 +37,8 @@ import androidx.compose.ui.unit.coerceAtMost
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import au.com.shiftyjelly.pocketcasts.account.R
+import au.com.shiftyjelly.pocketcasts.account.onboarding.components.ContinueWithGoogleButton
+import au.com.shiftyjelly.pocketcasts.account.viewmodel.GoogleSignInButtonViewModel
 import au.com.shiftyjelly.pocketcasts.account.viewmodel.GoogleSignInState
 import au.com.shiftyjelly.pocketcasts.account.viewmodel.OnboardingLoginOrSignUpViewModel
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
@@ -49,7 +46,6 @@ import au.com.shiftyjelly.pocketcasts.compose.CallOnce
 import au.com.shiftyjelly.pocketcasts.compose.bars.NavigationButton
 import au.com.shiftyjelly.pocketcasts.compose.bars.NavigationIconButton
 import au.com.shiftyjelly.pocketcasts.compose.buttons.RowButton
-import au.com.shiftyjelly.pocketcasts.compose.buttons.RowOutlinedButton
 import au.com.shiftyjelly.pocketcasts.compose.buttons.RowTextButton
 import au.com.shiftyjelly.pocketcasts.compose.components.PodcastCover
 import au.com.shiftyjelly.pocketcasts.compose.components.RectangleCover
@@ -63,7 +59,6 @@ import au.com.shiftyjelly.pocketcasts.settings.onboarding.OnboardingFlow
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.utils.extensions.pxToDp
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
-import au.com.shiftyjelly.pocketcasts.images.R as IR
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @Composable
@@ -131,7 +126,8 @@ internal fun OnboardingLoginOrSignUpPage(
 
         Spacer(Modifier.height(32.dp))
 
-        Artwork(viewModel.showContinueWithGoogleButton, viewModel.randomPodcasts)
+        val context = LocalContext.current
+        Artwork(GoogleSignInButtonViewModel.showContinueWithGoogleButton(context), viewModel.randomPodcasts)
 
         Spacer(Modifier.weight(1f))
 
@@ -159,7 +155,6 @@ internal fun OnboardingLoginOrSignUpPage(
             Spacer(Modifier.height(8.dp))
             ContinueWithGoogleButton(
                 flow = flow,
-                viewModel = viewModel,
                 onComplete = onContinueWithGoogleComplete
             )
         } else {
@@ -227,69 +222,6 @@ private fun Artwork(
             )
         }
     }
-}
-
-/**
- * Let the user sign into Pocket Casts with their Google account.
- * The One Tap for Android library is used. Sign in doesn't work when no Google accounts are set up on the device. In this case, fallback to the legacy Google Sign-In for Android.
- */
-@Composable
-private fun ContinueWithGoogleButton(
-    flow: OnboardingFlow,
-    viewModel: OnboardingLoginOrSignUpViewModel,
-    onComplete: (GoogleSignInState) -> Unit
-) {
-    val context = LocalContext.current
-    val errorMessage = stringResource(LR.string.onboarding_continue_with_google_error)
-
-    val showError = {
-        Toast.makeText(context, errorMessage, Toast.LENGTH_LONG).show()
-    }
-
-    // request legacy Google Sign-In and process the result
-    val googleLegacySignInLauncher = rememberLauncherForActivityResult(ActivityResultContracts.StartIntentSenderForResult()) { result ->
-        viewModel.onGoogleLegacySignInResult(
-            result = result,
-            onSuccess = onComplete,
-            onError = showError
-        )
-    }
-
-    // request Google One Tap Sign-In and process the result
-    val googleOneTapSignInLauncher = rememberLauncherForActivityResult(ActivityResultContracts.StartIntentSenderForResult()) { result ->
-        viewModel.onGoogleOneTapSignInResult(
-            result = result,
-            onSuccess = onComplete,
-            onError = {
-                viewModel.startGoogleLegacySignIn(
-                    onSuccess = { request -> googleLegacySignInLauncher.launch(request) },
-                    onError = showError
-                )
-            }
-        )
-    }
-
-    val onSignInClick = {
-        viewModel.startGoogleOneTapSignIn(
-            flow = flow,
-            onSuccess = { request -> googleOneTapSignInLauncher.launch(request) },
-            onError = {
-                viewModel.startGoogleLegacySignIn(
-                    onSuccess = { request -> googleLegacySignInLauncher.launch(request) },
-                    onError = showError
-                )
-            }
-        )
-    }
-
-    RowOutlinedButton(
-        text = stringResource(LR.string.onboarding_continue_with_google),
-        leadingIcon = painterResource(IR.drawable.google_g),
-        tintIcon = false,
-        border = BorderStroke(2.dp, MaterialTheme.theme.colors.primaryInteractive03),
-        colors = ButtonDefaults.outlinedButtonColors(contentColor = MaterialTheme.theme.colors.primaryText01),
-        onClick = onSignInClick
-    )
 }
 
 @Composable
