+++ modules/features/account/src/main/res/layout-car/fragment_account.xml
@@ -63,13 +63,24 @@
                 app:layout_constraintTop_toTopOf="parent"
                 app:src="@drawable/ic_create_new_account" />
 
+            <androidx.compose.ui.platform.ComposeView
+                android:id="@+id/btnContinueWithGoogle"
+                android:layout_width="592dp"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="@dimen/car_create_account_btn_top_margin"
+                android:layout_marginStart="16dp"
+                android:layout_marginEnd="16dp"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/lblSaveYourPodcasts" />
+
             <com.google.android.material.button.MaterialButton
                 android:id="@+id/btnCreate"
                 style="@style/MaterialButtonStyle"
                 android:layout_width="592dp"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="16dp"
-                android:layout_marginTop="@dimen/car_create_account_btn_top_margin"
+                android:layout_marginTop="@dimen/car_sign_in_btn_top_margin"
                 android:layout_marginEnd="16dp"
                 android:fontFamily="sans-serif-medium"
                 android:letterSpacing="0.02"
@@ -81,7 +92,7 @@
                 app:cornerRadius="8dp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/lblSaveYourPodcasts" />
+                app:layout_constraintTop_toBottomOf="@id/btnContinueWithGoogle" />
 
             <com.google.android.material.button.MaterialButton
                 android:id="@+id/btnSignIn"
